fix(plotter): cpu_threads u8 overflow on >=256-thread systems
256 as u8 == 0, so on dual-EPYC 7B12 (128c/256t) and similar high-core
systems rayon was handed a zero-thread pool and plotting silently never
progressed.

- pocx_plotter_v2: cpu_threads u8 -> usize in builder, task struct, and
  the conversion into PlotterTask; rayon pool now receives usize directly.
- pocx_plotter (legacy): same u8 -> usize migration through builder,
  PlotterTask, CLI parser, and the rayon pool build site.
- Tighten builder validation: explicit cpu_threads(0) is rejected even
  when a GPU is configured (previously silently built an inert CPU pool).
  GPU-only configs that leave cpu_threads at its default still build.
- Remove the stale 128-thread CLI cap in pocx_plotter; usize can express
  any real core count.
- Add unit tests covering 256/1024 thread counts, rayon pool sizing,
  explicit-zero rejection, and GPU-only build success.

Bump all workspace crates to 1.0.2.

+++ pocx_plotter/src/lib.rs
@@ -183,7 +183,7 @@ pub struct PlotterTaskBuilder {
     compress: u8,
     output_paths: Vec<String>,
     mem: String,
-    cpu_threads: u8,
+    cpu_threads: usize,
     gpus: Option<Vec<String>>,
     direct_io: bool,
     escalate: u64,
@@ -200,7 +200,7 @@ impl PlotterTaskBuilder {
     pub fn new() -> Self {
         Self {
             mem: "0B".to_string(), // Match CLI default format
-            cpu_threads: num_cpus::get() as u8,
+            cpu_threads: num_cpus::get(),
             direct_io: true,
             escalate: 1,
             quiet: true,
@@ -245,7 +245,7 @@ impl PlotterTaskBuilder {
         self
     }
 
-    pub fn cpu_threads(mut self, threads: u8) -> Self {
+    pub fn cpu_threads(mut self, threads: usize) -> Self {
         self.cpu_threads = threads;
         self
     }
